apps/reflex-chat/pyproject.toml
[project]
name = "reflex-chat"
version = "0.7.0" # x-release-please-version
description = "Reflex Chat Application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "albert-client",
    "pdf-context",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "reflex>=0.7.11",
]

[project.scripts]
reflex-chat = "reflex.reflex:cli"

[tool.setuptools.packages.find]
include = ["reflex_chat*"]

[tool.setuptools]
py-modules = ["context_loader"]

[tool.uv.sources]
albert-client = { workspace = true }
pdf-context = { workspace = true }
